backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1

